# server/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "DCRM API",
    "description": "API REST du CRM interne DCB Technologies. Format de réponse unifié : succès `{ success: true, data, meta? }`, erreur `{ success: false, error: { code, message } }`. Authentification par JWT Bearer (15 min, refresh 7 j via cookie httpOnly) ou par clé API longue durée (header X-API-Key). L'extension `x-permission` sur une opération indique la clé de permission RBAC requise. Tous les montants monétaires sont exprimés en euros hors taxes (HT) ; le TTC n'est jamais stocké et se dérive via Product.vatRate (défaut 20 %).",
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3001/api",
      "description": "Développement"
    },
    {
      "url": "https://{domaine}/api",
      "description": "Production",
      "variables": {
        "domaine": {
          "default": "crm.example.fr"
        }
      }
    }
  ],
  "security": [
    {
      "bearerAuth": []
    },
    {
      "apiKeyAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Access token obtenu via POST /auth/login (durée 15 min)"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Clé API longue durée générée via POST /apikeys (préfixe dcrm_). Permissions effectives = intersection des droits de la clé (scopes) et des droits actuels de son propriétaire — jamais de bypass '*', même pour un ADMIN ; les routes gardées par rôle sont inaccessibles par clé API."
      }
    },
...
  },
  "paths": {
...
    "/references": {
      "get": {
        "tags": [
          "Référentiels"
        ],
        "summary": "Tous les domaines de listes personnalisables et leurs valeurs (actives et désactivées)",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "data: [{ domain, label, description, validate, keyStyle, hasColor, hasIcon, values: [{ id, key, label, color, icon, order, isActive, isSystem, meta }] }]"
          }
        }
      }
    },
    "/references/{domain}": {
      "post": {
        "tags": [
          "Référentiels"
        ],
        "summary": "Ajoute une valeur à un domaine (clé générée du libellé si absente)",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ],
        "x-permission": "references:manage",
        "parameters": [
          {
            "name": "domain",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "enum": ["ticket_category", "call_category", "equipment_type", "equipment_status", "contract_type", "license_type", "contact_status", "lead_source", "sector", "appointment_type", "knowledge_category", "product_category"]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "label"
                ],
                "properties": {
                  "label": {
                    "type": "string",
                    "maxLength": 80
                  },
                  "key": {
                    "type": "string",
                    "maxLength": 80
                  },
                  "color": {
                    "type": "string",
                    "enum": ["gray", "slate", "blue", "indigo", "violet", "purple", "pink", "red", "orange", "amber", "yellow", "green", "emerald", "cyan"],
                    "nullable": true
                  },
                  "icon": {
                    "type": "string",
                    "maxLength": 40,
                    "nullable": true
                  },
                  "meta": {
                    "type": "object",
                    "nullable": true,
                    "description": "JSON libre — ex. { \"isPhysical\": true } sur product_category"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "ReferenceValue créée"
          },
          "400": {
            "description": "INVALID_KEY / VALIDATION_ERROR"
          },
          "404": {
            "description": "NOT_FOUND — domaine inconnu"
          },
          "409": {
            "description": "DUPLICATE — clé déjà présente dans le domaine"
          }
        }
      }
    },
    "/references/{id}": {
      "put": {
        "tags": [
          "Référentiels"
        ],
        "summary": "Modifie libellé/couleur/icône/meta/actif d'une valeur (clé immuable ; valeurs système non désactivables)",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ],
        "x-permission": "references:manage",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "label": {
                    "type": "string",
                    "maxLength": 80
                  },
                  "color": {
                    "type": "string",
                    "nullable": true
                  },
                  "icon": {
                    "type": "string",
                    "nullable": true
                  },
                  "meta": {
                    "type": "object",
                    "nullable": true
                  },
                  "isActive": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "ReferenceValue mise à jour"
          },
          "400": {
            "description": "SYSTEM_VALUE — valeur système non désactivable / VALIDATION_ERROR"
          },
          "404": {
            "description": "NOT_FOUND"
          }
        }
      },
      "delete": {
        "tags": [
          "Référentiels"
        ],
        "summary": "Supprime une valeur — ou la désactive si elle est utilisée par des données existantes",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ],
        "x-permission": "references:manage",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "{ deleted: true } ou { deactivated: true, usage: n }"
          },
          "400": {
            "description": "SYSTEM_VALUE — valeur système non supprimable"
          },
          "404": {
            "description": "NOT_FOUND"
          }
        }
      }
    },
    "/references/{domain}/reorder": {
      "patch": {
        "tags": [
          "Référentiels"
        ],
        "summary": "Réordonne les valeurs d'un domaine",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ],
        "x-permission": "references:manage",
        "parameters": [
          {
            "name": "domain",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "ids"
                ],
                "properties": {
                  "ids": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "minItems": 1
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "{ reordered: n }"
          },
          "404": {
            "description": "NOT_FOUND — domaine inconnu"
          }
        }
      }
    },
...
  }
}